Add files via upload
This is a calendar slot booking service app, made using Tkinter (Python) it uses SQLite3 Database for fetching the details of team members for scheduling the meetings. Proper validation is done in this app. The screenshots are attached with the application calender.py Some modules have to be imported for running this program like Pillow, tkcalendar.